+++ tests/files/webpack.config.js
@@ -2,5 +2,8 @@ module.exports = {
   resolve: {
     extensions: ['', '.js', '.jsx'],
     root: __dirname,
+    alias: {
+      'alias/chai$': 'chai' // alias for no-extraneous-dependencies tests
+    }
   },
 }