+++ tests/src/rules/order.js
@@ -376,23 +376,6 @@ ruleTester.run('order', rule, {
         'import/external-module-folders': ['node_modules', 'symlinked-module'],
       },
     }),
-    // Monorepo setup, using Webpack resolver, partial workspace folder path
-    // in external-module-folders
-    test({
-      code: `
-        import _ from 'lodash';
-        import m from '@test-scope/some-module';
-
-        import bar from './bar';
-      `,
-      options: [{
-        'newlines-between': 'always',
-      }],
-      settings: {
-        'import/resolver': 'webpack',
-        'import/external-module-folders': ['node_modules', 'files/symlinked-module'],
-      },
-    }),
     // Monorepo setup, using Node resolver (doesn't resolve symlinks)
     test({
       code: `
@@ -406,7 +389,7 @@ ruleTester.run('order', rule, {
       }],
       settings: {
         'import/resolver': 'node',
-        'import/external-module-folders': ['node_modules', 'files/symlinked-module'],
+        'import/external-module-folders': ['node_modules', 'symlinked-module'],
       },
     }),
     // Option: newlines-between: 'always'